+++ .gitignore
@@ -1,6 +1,19 @@
+### Gradle ###
+.gradle
+build/
+!gradle/wrapper/gradle-wrapper.jar
+!**/src/main/**/build/
+!**/src/test/**/build/
+
 ### IntelliJ IDEA ###
-out/production
-!out/artifacts/CMRITLeaderboard2025_jar/*
+.idea/modules.xml
+.idea/jarRepositories.xml
+.idea/compiler.xml
+.idea/libraries/
+*.iws
+*.iml
+*.ipr
+out/
 !**/src/main/**/out/
 !**/src/test/**/out/
