--- .idea/artifacts/CMRITLeaderboard2025_jar.xml
@@ -1,18 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="CMRITLeaderboard2025:jar">
-    <output-path>$PROJECT_DIR$/out/artifacts/CMRITLeaderboard2025_jar</output-path>
-    <root id="archive" name="CMRITLeaderboard2025.jar">
-      <element id="module-output" name="CMRITLeaderboard2025" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/java-json.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/poi-4.0.1.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/curvesapi-1.05.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/xmlbeans-3.0.2.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/poi-ooxml-4.0.1.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/commons-codec-1.11.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/commons-math3-3.6.1.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/commons-compress-1.18.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/poi-ooxml-schemas-4.0.1.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/commons-collections4-4.2.jar" path-in-jar="/" />
-    </root>
-  </artifact>
-</component>
