+++ src/main/resources/index.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+
+<html lang="en">
+
+<head >
+<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=0">
+<link rel="shortcut icon" type="image/x-icon" href="https://github.com/gabyah92/CMRIT2025Leaderboard/blob/main/lib/logonobg.png?raw=true">
+    <title>CMRIT 2025 Leaderboard</title>
+    
+  <meta charSet="UTF-8" /> 
+    <link href="https://fonts.googleapis.com/css2?family=IBM+Plex+Sans:wght@400;500;700&amp;display=swap"
+        rel="stylesheet" />
+    <style media="only screen">
+        html,
+        body {
+            height: 100%;
+            width: 100%;
+            margin: 0;
+            box-sizing: border-box;
+            -webkit-overflow-scrolling: touch;
+        }
+
+        html {
+            position: absolute;
+            top: 0;
+            left: 0;
+            padding: 0;
+            overflow: auto;
+        }
+
+        body {
+            padding: 25px;
+            overflow: auto;
+            background-color: black
+        }
+    </style>
+    <link rel="stylesheet" href="styles.css" />
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/1.12.1/jquery.min.js">
+    </script>
+    <script src="https://cdn.jsdelivr.net/npm/xlsx-style@0.8.13/dist/xlsx.full.min.js">
+    </script>
+</head>
+
+<body>
+    <div class="example-wrapper">
+        <div id="myGrid" class="ag-theme-quartz-dark">
+        </div>
+    </div>
+    <script>var __basePath = './';</script>
+    <script src="https://cdn.jsdelivr.net/npm/ag-grid-community/dist/ag-grid-community.min.js"></script>
+    <script src="app.js">
+    </script>
+</body>
+<link rel="stylesheet" href="styles.css" />
+<h1>Excel Sheet: <a href="https://github.com/gabyah92/CMRIT2025Leaderboard/raw/main/Leaderboards/CurrentCMRITLeaderboard2025.xlsx"> Download</a> </h1>
+<h1>App By : <a href="https://www.instagram.com/gabyah92">gabyah92</a> & <a href="https://github.com/dog-broad">Rushyendra</a>  </h1>
+</html>
