+++ src/main/resources/styles.css
@@ -1,0 +1,21 @@
+.example-wrapper {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+}
+h1{ 
+  color:white;
+}
+a{
+  color: orange;
+}
+#myGrid {
+    flex: 1 1 0px;
+    width: 100%;
+} 
+div.ag-root .ag-cell-focus {
+    -webkit-user-select: text;
+    -moz-user-select: text;
+    -ms-user-select: text;
+    user-select: text;
+}
